+++ lib/ooxml_parser/common_parser/common_data/alternate_content/chart/chart/plot_area/common_chart_data.rb
@@ -1,0 +1,29 @@
+module OoxmlParser
+  # Parsing common chart data
+  class CommonChartData < OOXMLDocumentObject
+    # @return [ValuedChild] type of grouping
+    attr_reader :grouping
+    # @return [Array<Series>] series data
+    attr_reader :series
+
+    def initialize(parent: nil)
+      @series = []
+      @parent = parent
+    end
+
+    # Parse View3D object
+    # @param node [Nokogiri::XML:Element] node to parse
+    # @return [View3D] result of parsing
+    def parse(node)
+      node.xpath('*').each do |node_child|
+        case node_child.name
+        when 'grouping'
+          @grouping = ValuedChild.new(:symbol, parent: self).parse(node_child)
+        when 'ser'
+          @series << Series.new(parent: self).parse(node_child)
+        end
+      end
+      self
+    end
+  end
+end